Move `ty::Const` and `ty::ConstKind` into their own modules
